# Fix performance regression on crate search