--- a/migrations/2019-05-14-165316_index_crate_name_for_like/down.sql
+++ b/migrations/2019-05-14-165316_index_crate_name_for_like/down.sql
@@ -1,0 +1,2 @@
+DROP INDEX index_crates_name_tgrm;
+DROP EXTENSION pg_trgm;