--- a/migrations/2019-05-14-165316_index_crate_name_for_like/up.sql
+++ b/migrations/2019-05-14-165316_index_crate_name_for_like/up.sql
@@ -1,0 +1,2 @@
+CREATE EXTENSION pg_trgm;
+CREATE INDEX index_crates_name_tgrm ON crates USING gin (canon_crate_name(name) gin_trgm_ops);